user & admin access

+++ dashboard/views.py
@@ -3,15 +3,19 @@ from django.http import HttpResponse
 from dashboard.forms import SlidImageForm
 from dashboard.models import Image
 from post.models import Blog, Comment
+from django.contrib.auth.decorators import login_required
 
-
+@login_required(login_url='login')
 def dashboard1(request):
-
+    if not request.user.is_staff:
+        return HttpResponse("Permission denied")
     # return render(request,'Dashboard/base.html')
-
     return render(request, 'dashboard/base.html')
 
+@login_required(login_url='login')
 def image_upload(request):
+    if not request.user.is_staff:
+        return HttpResponse("Permission denied")
     if request.method == 'POST':
         form = SlidImageForm(request.POST, request.FILES)
         if form.is_valid():
@@ -29,8 +33,10 @@ def image_upload(request):
         return render (request, 'dashboard/upload_image/image_upload.html',context)       
 
     return render (request, 'dashboard/upload_image/image_upload.html')    
-
+@login_required(login_url='login')
 def image_list(request):
+    if not request.user.is_staff:
+        return HttpResponse("Permission denied")
     if request.method == 'GET':
         photo_list = Image.objects.all()
 
@@ -38,14 +44,18 @@ def image_list(request):
             'photo_list' : photo_list
         }
     return render(request,'dashboard/upload_image/image_list.html', context)
-
+@login_required(login_url='login')
 def delete_image(request, id):
+    if not request.user.is_staff:
+        return HttpResponse("Permission denied")
     if request.method == 'POST':
         image_list = Image.objects.get(id = id)
         image_list.delete()
         return redirect('image_list')
-
+@login_required(login_url='login')
 def manage_post(request):
+    if not request.user.is_staff:
+        return HttpResponse("Permission denied")
     post = Blog.objects.all()
     comment  = Comment.objects.all()
 
@@ -55,14 +65,18 @@ def manage_post(request):
         
     }
     return render(request, 'dashboard/manage_post/manage_post.html', context)
-
+@login_required(login_url='login')    
 def delete_post(request,id):
+    if not request.user.is_staff:
+        return HttpResponse("Permission denied")
     if request.method == 'POST':
         post = Blog.objects.get(id = id)
         post.delete()
         return redirect('manage_post')
-
+@login_required(login_url='login')
 def individual_post(request, id):
+    if not request.user.is_staff:
+        return HttpResponse("Permission denied")
     if request.method == 'GET':
         post_item = Blog.objects.filter(id = id)
         comments = Comment.objects.filter(blog__id = id).all()
@@ -73,8 +87,10 @@ def individual_post(request, id):
         }
     return render(request, 'dashboard/manage_post/individual_post.html', context)
 
-
+@login_required(login_url='login')
 def delete_comment(request, id):
+    if not request.user.is_staff:
+        return HttpResponse("Permission denied")
     comment = Comment.objects.filter(id = id).first()
     comment.delete()
